Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0119-pascals-triangle-ii/0119-pascals-triangle-ii.java
@@ -1,12 +1,13 @@
 class Solution {
     public List<Integer> getRow(int rowIndex) {
         List<Integer> list = new ArrayList<>();
-        long num = 1;
+        int num = 1;
         
         for (int i = 0; i <= rowIndex; i++) {
-            list.add((int)num);
+            list.add(num);
             num = num * (rowIndex - i) / (i + 1);
         }
+
         return list;
     }
 }
